+++ src/main/kotlin/com/emberjs/hbs/HbsLocalReference.kt
@@ -145,9 +145,16 @@ class HbsLocalReference(private val leaf: PsiElement, val resolved: Any?) : HbRe
             if (parts.first() == "this") {
                 current = JSContextResolver.resolveThisReference(tpl as PsiElement)
             } else {
-                val children = PsiTreeUtil.collectElements(f) { it is JSVariable || it is ES6ImportDeclaration }
+                val children = PsiTreeUtil.collectElements(f) { it is JSVariable || it is ES6ImportDeclaration || it is JSClass }
                 current = children.mapNotNull {
                     if (it is JSVariable && it.name?.equals(parts.first()) == true) {
+                        val useScope = JSUseScopeProvider.getBlockScopeElement(it)
+                        if (useScope.isAncestor(tpl as PsiElement)) {
+                            return@mapNotNull it
+                        }
+                    }
+
+                    if (it is JSClass && it.name?.equals(parts.first()) == true) {
                         val useScope = JSUseScopeProvider.getBlockScopeElement(it)
                         if (useScope.isAncestor(tpl as PsiElement)) {
                             return@mapNotNull it